requirements.txt
discord.py
Pyrebase4